#39 Rename shared as interface adapter and fix imports
