+++ tests/interface_adapters/test_request_object.py
@@ -1,4 +1,4 @@
-from webminer.shared import request_object as req
+from webminer.interface_adapters import request_object as req
 
 
 def test_invalid_request_object_is_false():
