+++ webminer/domain/arxiv_document.py
@@ -4,7 +4,7 @@ Returns:
     class: ArxivDocument class
 """
 
-from webminer.shared.domain_model import DomainModel
+from webminer.interface_adapters.domain_model import DomainModel
 
 
 class ArxivDocument(object):
