+++ webminer/interface_adapters/domain_model.py
