+++ webminer/interface_adapters/response_object.py
