+++ webminer/use_cases/arxiv_document_use_case.py
@@ -4,8 +4,8 @@ Returns:
     class: For processing arxiv document list and apply the filtering on a repo
 """
 
-from webminer.shared import use_case as uc
-from webminer.shared import response_object as res
+from webminer.interface_adapters import use_case as uc
+from webminer.interface_adapters import response_object as res
 
 
 class ArxivDocumentListUseCase(uc.UseCase):
