+++ webminer/use_cases/request_objects.py
@@ -1,7 +1,7 @@
 """Takes the valid request object and returns a filtered arxiv document"""
 
 import collections
-from webminer.shared import request_object as req
+from webminer.interface_adapters import request_object as req
 
 
 class ArxivDocumentListRequestObject(req.ValidRequestObject):
